Topic/build 60

+++ libNTService/NTServiceGetopt.h
@@ -1,10 +1,11 @@
 #pragma once
 #ifndef NTSERVICEGETOPT_H_INCLUDED
 #define NTSERVICEGETOPT_H_INCLUDED
+
 /*
  * CNTService - Classic window services framework (tweaked).
  *
- * Copyright (c) 2020, Adam Young.
+ * Copyright (c) 2020 - 2022, Adam Young.
  * All rights reserved.
  *
  * This file is part of the WinRSH/WinSSH project.
@@ -24,7 +25,7 @@
  * This project is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * License for more details.
+ * license for more details.
  * ==end==
  */
 
@@ -49,6 +50,7 @@
 #include <cstdio>
 #include <cstdlib>
 #include <cstring>
+#include <iostream>
 #include <cassert>
 
 #include "NTServiceIO.h"
@@ -241,7 +243,7 @@ private:
                 place_ = "";
 
                 if ((has_equal = strchr(arg, '=')) != NULL) {
-                        arglen = has_equal - arg;
+                        arglen = (int)(has_equal - arg);
                         ++has_equal;
                 } else {
                         arglen = (int)strlen(arg);
@@ -327,7 +329,7 @@ private:
                 if (!opterr_) return;
 
                 char buffer[1024];
-                if (arglen < 0) arglen = strlen(arg);
+                if (arglen < 0) arglen = (int)strlen(arg);
                 (void) _snprintf(buffer, sizeof(buffer), "%s: %s -- %.*s", progname(), msg, arglen, arg);
                 buffer[sizeof(buffer) - 1] = 0;
                 error_report(code, buffer);
@@ -351,5 +353,3 @@ private:
 }   //namspace NTService
 
 #endif  NTSERVICEGETOPT_HPP_INCLUDED
-
-
